+++ WooCommerce/Classes/POS/Presentation/PointOfSaleDashboardView.swift
@@ -11,18 +11,45 @@ struct PointOfSaleDashboardView: View {
 
     var body: some View {
         VStack {
-            Text("WooCommerce Point Of Sale")
-                .foregroundColor(Color.white)
             HStack {
                 switch viewModel.orderStage {
                 case .building:
                     productGridView
                     Spacer()
-                    cartView
+                    if viewModel.isCartCollapsed {
+                        collapsedCartView
+                    } else {
+                        cartView
+                    }
                 case .finalizing:
                     cartView
                     Spacer()
-                    totalsView
+                    VStack {
+                        totalsView
+                        // TODO: replace temporary inline message UI based on design
+                        if let inlinePaymentMessage = viewModel.cardPresentPaymentInlineMessage {
+                            switch inlinePaymentMessage {
+                            case .preparingForPayment:
+                                Text("Preparing for payment...")
+                            case .tapSwipeOrInsertCard:
+                                Text("tapSwipeOrInsertCard...")
+                            case .processing:
+                                Text("processing...")
+                            case .displayReaderMessage(let message):
+                                Text("Reader message: \(message)")
+                            case .success:
+                                Text("Payment successful!")
+                            case .error:
+                                Text("Payment error")
+                            case .nonRetryableError:
+                                Text("Payment error - non retryable")
+                            case .cancelledOnReader:
+                                Text("Payment cancelled on reader")
+                            }
+                        }
+                    }
+                    // TODO: remove this after replacing temporary inline message UI based on design
+                    .background(Color.orange)
                 }
             }
             .padding()
@@ -30,38 +57,41 @@ struct PointOfSaleDashboardView: View {
         .background(Color.primaryBackground)
         .navigationBarBackButtonHidden(true)
         .toolbar {
-            ToolbarItem(placement: .topBarLeading, content: {
-                Button("Exit POS") {
-                    presentationMode.wrappedValue.dismiss()
-                }
-            })
-            ToolbarItem(placement: .principal, content: {
-                CardReaderConnectionStatusView(connectionViewModel: viewModel.cardReaderConnectionViewModel)
-            })
-            ToolbarItem(placement: .primaryAction, content: {
-                Button("History") {
-                    debugPrint("Not implemented")
-                }
-            })
-        }
-        .sheet(isPresented: $viewModel.showsCardReaderSheet, content: {
-            switch viewModel.cardPresentPaymentEvent {
-            case .showAlert(let alertViewModel):
-                CardPresentPaymentAlert(alertViewModel: alertViewModel)
-            case .idle,
-                    .showReaderList,
-                    .showOnboarding:
-                Text(viewModel.cardPresentPaymentEvent.temporaryEventDescription)
+            ToolbarItem(placement: .bottomBar) {
+                POSToolbarView(readerConnectionViewModel: viewModel.cardReaderConnectionViewModel)
             }
-        })
-        .sheet(isPresented: $viewModel.showsFilterSheet, content: {
-            FilterView(viewModel: viewModel)
+        }
+        .toolbarBackground(Color.toolbarBackground, for: .bottomBar)
+        .toolbarBackground(.visible, for: .bottomBar)
+        .sheet(isPresented: $viewModel.showsCardReaderSheet, content: {
+            // Might be the only way unless we make the type conform to `Identifiable`
+            if let alertType = viewModel.cardPresentPaymentAlertViewModel {
+                PointOfSaleCardPresentPaymentAlert(alertType: alertType)
+            } else {
+                switch viewModel.cardPresentPaymentEvent {
+                case let .showReaderList(readerIDs, selectionHandler):
+                    // TODO: make this an instance of `showAlert` so we can handle it above too.
+                    FoundCardReaderListView(readerIDs: readerIDs, connect: { readerID in
+                        selectionHandler(readerID)
+                    }, cancelSearch: {
+                        selectionHandler(nil)
+                    })
+                case .idle,
+                        .show, // handled above
+                        .showOnboarding:
+                    Text(viewModel.cardPresentPaymentEvent.temporaryEventDescription)
+                }
+            }
         })
     }
 }
 
 /// Helpers to generate all Dashboard subviews
 private extension PointOfSaleDashboardView {
+    var collapsedCartView: some View {
+        CollapsedCartView()
+    }
+
     var cartView: some View {
         CartView(viewModel: viewModel)
             .background(Color.secondaryBackground)
@@ -75,7 +105,7 @@ private extension PointOfSaleDashboardView {
     }
 
     var productGridView: some View {
-        ItemGridView(viewModel: viewModel)
+        ItemListView(viewModel: viewModel)
             .background(Color.secondaryBackground)
             .frame(maxWidth: .infinity)
     }
@@ -86,8 +116,8 @@ fileprivate extension CardPresentPaymentEvent {
         switch self {
         case .idle:
             return "Idle"
-        case .showAlert(let alertViewModel):
-            return "Alert: \(alertViewModel.topTitle)"
+        case .show:
+            return "Event"
         case .showReaderList(let readerIDs, _):
             return "Reader List: \(readerIDs.joined())"
         case .showOnboarding(let onboardingViewModel):
@@ -98,8 +128,10 @@ fileprivate extension CardPresentPaymentEvent {
 
 #if DEBUG
 #Preview {
-    PointOfSaleDashboardView(
-        viewModel: PointOfSaleDashboardViewModel(items: POSItemProviderPreview().providePointOfSaleItems(),
-                                                 cardPresentPaymentService: CardPresentPaymentPreviewService()))
+    NavigationStack {
+        PointOfSaleDashboardView(
+            viewModel: PointOfSaleDashboardViewModel(items: POSItemProviderPreview().providePointOfSaleItems(),
+                                                     cardPresentPaymentService: CardPresentPaymentPreviewService()))
+    }
 }
 #endif