+++ WooCommerce/Classes/ViewRelated/Dashboard/DashboardView.swift
@@ -107,7 +107,8 @@ struct DashboardView: View {
                 }, label: {
                     Text(Localization.edit)
                         .overlay(alignment: .topTrailing) {
-                            if viewModel.showNewCardsNotice {
+                            if viewModel.showNewCardsNotice &&
+                                !viewModel.isReloadingAllData {
                                 Circle()
                                     .fill(Color(.accent))
                                     .frame(width: Layout.dotBadgeSize)
@@ -131,10 +132,7 @@ struct DashboardView: View {
             connectivityStatus = status
         }
         .refreshable {
-            Task { @MainActor in
-                ServiceLocator.analytics.track(.dashboardPulledToRefresh)
-                await viewModel.reloadAllData()
-            }
+            viewModel.onPullToRefresh()
         }
         .safeAreaInset(edge: .bottom) {
             jetpackBenefitBanner
@@ -155,9 +153,7 @@ struct DashboardView: View {
         }
         .sheet(isPresented: $viewModel.showingCustomization,
                onDismiss: {
-            Task {
-                await viewModel.handleCustomizationDismissal()
-            }
+            viewModel.handleCustomizationDismissal()
         }) {
             DashboardCustomizationView(viewModel: DashboardCustomizationViewModel(
                 allCards: viewModel.availableCards,
@@ -169,7 +165,9 @@ struct DashboardView: View {
             Survey(source: .inAppFeedback)
         }
         .onAppear {
-            viewModel.onViewAppear()
+            Task {
+                await viewModel.onViewAppear()
+            }
         }
     }
 }
@@ -246,11 +244,11 @@ private extension DashboardView {
                 }
             }
 
-            if viewModel.showNewCardsNotice {
+            if viewModel.showNewCardsNotice && !viewModel.isReloadingAllData {
                 newCardsNoticeCard
             }
 
-            if !viewModel.hasOrders {
+            if !viewModel.hasOrders && !viewModel.isReloadingAllData {
                 shareStoreCard
             }
         }
@@ -307,14 +305,14 @@ private extension DashboardView {
                 .multilineTextAlignment(.center)
                 .padding(.horizontal, Layout.elementPadding)
 
-            Button(Localization.NewCardsNoticeCard.addSectionsButtonLabel) {
+            Button(Localization.NewCardsNoticeCard.addSectionsButtonText) {
                 ServiceLocator.analytics.track(event: .DynamicDashboard.dashboardCardAddNewSectionsTapped())
 
                 viewModel.showCustomizationScreen()
             }
             .buttonStyle(PrimaryButtonStyle())
             .padding(.horizontal, Layout.elementPadding)
-            .padding(.bottom, Layout.padding)
+            .padding(.bottom, Layout.elementPadding)
         }
         .background(Color(.listForeground(modal: false)))
         .clipShape(RoundedRectangle(cornerSize: Layout.cornerSize))
@@ -439,9 +437,9 @@ private extension DashboardView {
                 comment: "Subtitle of the New Cards Notice card"
             )
 
-            static let addSectionsButtonLabel = NSLocalizedString(
-                "dashboardView.newCardsNoticeCard.addSectionsButtonLabel",
-                value: "Add new sections",
+            static let addSectionsButtonText = NSLocalizedString(
+                "dashboardView.newCardsNoticeCard.addSectionsButtonText",
+                value: "Add New Sections",
                 comment: "Label of the button to add sections"
             )
         }